We do not really need a int value here

+++ app/reports/SimpleAssetReport.scala
@@ -14,7 +14,7 @@ class SimpleAssetReport @Inject() (protected val rs: RateService, protected val 
                                   (implicit ec: SqlExecutionContext){
   def calculate(start: LocalDate, end: LocalDate, granularity: Int): Future[Seq[(LocalDate, BigDecimal)]] = {
     val numberOfDays = ChronoUnit.DAYS.between(start, end)/granularity
-    val daysGenerated = for (f <- 0 to numberOfDays.intValue()) yield start.plusDays(f*granularity)
+    val daysGenerated = for (f <- 0L to numberOfDays) yield start.plusDays(f*granularity)
     val days = daysGenerated.:+(end)
 
     val series = days.map(d => sql.query(BudgetQuery.getTotalAssetsForDate(d)).map((d,_)))
